int32 changes

+++ api/middleware.go
@@ -20,7 +20,7 @@ func (m *Middleware) ValidateStagedFileExists(c *fiber.Ctx) error {
 	res, err := m.AtlasQueryClients.Storage.File(c.Context(), &req)
 	if err != nil {
 		return c.Status(400).SendString("Unable to find staged file with ID " + fileId)
-	} else if int64(len(res.File.Providers)) >= res.File.Replicas {
+	} else if len(res.File.Providers) >= res.File.Replicas {
 		return c.Status(409).SendString("File with ID " + fileId + "already has max. replicas stored accross the network.")
 	} else {
 		return c.Next()
